+++ CHANGES.md
@@ -1,6 +1,7 @@
 # RegReplace
 
-- **FIX**:  Properly set default replace to group 0 by using `\g<0>` not `\0`.
+- **NEW**: Add support for optional format style replace templates. Learn more [here](http://facelessuser.github.io/backrefs/#format-replacements).
+- **FIX**: Properly set default replace to group 0 by using `\g<0>` not `\0`.
 
 # Regreplace 3.5.0